Splited registration form and editing of house
