+++ src/main/java/ru/geekbrains/gkportal/controllers/AdminController.java
@@ -3,14 +3,16 @@ package ru.geekbrains.gkportal.controllers;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
-import org.springframework.web.bind.annotation.*;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.RequestMapping;
 import ru.geekbrains.gkportal.DTO.House;
-import ru.geekbrains.gkportal.DTO.HouseSimple;
 import ru.geekbrains.gkportal.DTO.SystemUser;
 import ru.geekbrains.gkportal.services.HouseService;
 
 @Controller
-public class RegistrationController {
+@RequestMapping("/admin")
+public class AdminController {
 
     private HouseService houseService;
 
@@ -24,14 +26,6 @@ public class RegistrationController {
         House house = houseService.build(number);
         model.addAttribute("house", house);
         model.addAttribute("systemUser", new SystemUser());
-        return "reg-form";
-    }
-
-
-    @PostMapping("/userRegister")
-    @ResponseBody
-    public String registerUser(@ModelAttribute("systemUser") SystemUser systemUser, Model model){
-        return systemUser.toString();
-
+        return "edit-house-form";
     }
 }
