+++ src/main/resources/templates/edit-house-form.html
@@ -6,7 +6,7 @@
     <script src="https://ajax.googleapis.com/ajax/libs/jquery/3.3.1/jquery.min.js"></script>
     <link rel="stylesheet" href="https://ajax.googleapis.com/ajax/libs/jqueryui/1.12.1/themes/smoothness/jquery-ui.css">
     <script src="https://ajax.googleapis.com/ajax/libs/jqueryui/1.12.1/jquery-ui.min.js"></script>
-    <script th:src="@{/js/regboard.js}" type="text/javascript"></script>
+    <script th:src="@{/js/editboard.js}" type="text/javascript"></script>
     <title>Title</title>
     <style>
         .porch {
@@ -58,22 +58,10 @@
 
     </table>
 </div>
-
+<div><button id="edit">Редактирование</button>
+<button id="stopedit">Перестать редактировать</button></div>
 <div>
-    <form th:action=" @{
-                '/userRegister'}" th:object="${systemUser}" method="post">
-                <div><label>Фамилия<input th:field="*{lastName}" placeholder="Иванов"></label>
-                    <div></div>
-                </div>
-                <div><label>Имя<input th:field="*{firstName}" placeholder="Иван"></label></div>
-                <div><label>Отчество<input th:field="*{middleName}" placeholder="Иванович"></label></div>
-                <div><label>Номер телефона<input th:field="*{phoneNumber}" placeholder="+7 999 123-45-67"></label></div>
-                <div><label>Адрес электронной почты<input type="email" th:field="*{email}"></label></div>
-                <div><label>Пароль </label></div>
-                <div>
-                    <button type="submit">Отправить</button>
-                </div>
-                </form>
+
 </div>
 </body>
 </html>
